request and login page

+++ grails-app/services/cashmanagement/LoanService.groovy
@@ -6,7 +6,7 @@ class LoanService {
         (new Random().nextInt() + 10000000).toString()
     }
 
-    def checkResourceAvailability(Branch branch, long amt) {
+    def checkResourceAvailability(Branch branch, double amt) {
         Math.random() > 0.5
     }
 }
